Create authQueries file
