+++ src/components/Auth/useAuthProvider.ts
@@ -1,37 +1,8 @@
 import { useClient } from '@vocdoni/react-providers'
 import { RemoteSigner } from '@vocdoni/sdk'
 import { useCallback, useEffect, useMemo, useState } from 'react'
-import { useMutation, UseMutationOptions } from '@tanstack/react-query'
 import { api, ApiEndpoints, ApiParams, UnauthorizedApiError } from '~components/Auth/api'
-
-export type LoginResponse = { token: string; expirity: string }
-
-export interface ILoginParams {
-  email: string
-  password: string
-}
-
-export interface IRegisterParams {
-  firstName: string
-  lastName: string
-  email: string
-  password: string
-}
-
-export const useLogin = (options: Omit<UseMutationOptions<LoginResponse, Error, ILoginParams>, 'mutationFn'>) => {
-  return useMutation<LoginResponse, Error, ILoginParams>({
-    mutationFn: (params: ILoginParams) => api<LoginResponse>(ApiEndpoints.LOGIN, { body: params, method: 'POST' }),
-    ...options,
-  })
-}
-
-export const useRegister = (options: Omit<UseMutationOptions<LoginResponse, Error, IRegisterParams>, 'mutationFn'>) => {
-  return useMutation<LoginResponse, Error, IRegisterParams>({
-    mutationFn: (params: IRegisterParams) =>
-      api<LoginResponse>(ApiEndpoints.REGISTER, { body: params, method: 'POST' }),
-    ...options,
-  })
-}
+import { LoginResponse, useLogin, useRegister } from '~components/Auth/authQueries'
 
 enum LocalStorageKeys {
   AUTH_TOKEN = 'authToken',
